requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.3
django-cors-headers==3.11.0
django-createsuperuserwithpassword==2.0.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
filelock==3.6.0
fonttools==4.31.2
gitignore==0.0.8
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.0
kiwisolver==1.4.2
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.1
protobuf==4.21.2
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
static3==0.7.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.13.3
whitenoise==6.0.0
